feat(moka): 本地 PG → Moka 开放平台部门全量推送
- IMokaDepartmentPushManager + Impl: 查 PG 全量 → DO→API DTO 映射 → 调 Moka PUT /v2/departments
- MokaDepartmentDO.sequence Integer → Moka API BigDecimal 转换
- MokaDepartmentLocalizedDO → MokaDepartment.localizedNames 多语言填充
- Controller 新增 POST /api/v1/moka/departments/push-to-moka
- 推送成功 moka_sync_status 置 1(SYNCED)，失败置 2(SYNC_FAILED)
- 全量推送语义：Moka API 以 departmentCode 为主键，必须推送本地全量

+++ cacch-integration-integration/src/main/java/com/cacch/integration/integration/moka/client/dto/MokaDepartment.java
@@ -2,7 +2,6 @@ package com.cacch.integration.integration.moka.client.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -39,9 +38,9 @@ public class MokaDepartment {
     private Integer type;
 
     /**
-     * 部门排序，范围 0~10000 支持两位小数；为空默认排在最后，按排序从小到大、创建时间从先到后（可选）
+     * 部门排序，范围 0~10000；为空默认排在最后，按排序从小到大、创建时间从先到后（可选）
      */
-    private BigDecimal sequence;
+    private Integer sequence;
 
     /**
      * 部门名称的多语言信息（可选；只能传入 Moka 已开通的语言；name 字段默认作为 zh-CN）
